requirements.txt
fastapi==0.99.1
pydantic==1.10.12
uvicorn==0.23.2
langchain==0.0.350
langserve==0.0.46
httpx==0.24.1
langchain-ollama
streamlit
python-dotenv
langchain-core
langchain-openai
langchain-community
sse-starlette


